Revert krill migration
Migrating kill to app does not seem to change the behavior of aquarium, but it still doesnt let you edit krill code live, all krill symbols disapear from the namespace as soon as a change is saved. No reason to bother with making this change unless we can get the benefits out of it.