Minor tweak to guard
This should usually be redundant, because assignments to one field value of a given routing symbol should always be the same for other field values with the same routing symbol. But I encountered a stiutation while debugging something where the sample types associated with two field values with the same routing symbol did not match, in which case the backend would un-assign the offending field values, leaving some assigned and some not.