Fix another test op generation edge case
Previously, routing an array of inputs to an array of outputs would not function. I have fixed this essentially with a bandaid, since the reason it didn't work in the first place is due to another error in the generation code which I cant find the source of. The pre existing error basically causes a generated output array to be triplicated, so taking the first set of outputs from val works to generate 3 valid outputs as expected, rather than 9.