Fix selection box glitch
Previously, when selecting the bottom row of part associations on a collection, the selection box would go outside the bottom bound of main-scrollbar, causing a scrollbar to appear on the right and the selection boxes to all be off by the width of a scrollbar. This margin ensures that the bottommost selection boxes will not overflow main-scrollbar.